refactor(types): split auth.ts into 5 per-entity files
